+++ lightningd/gossip_msg.h
@@ -6,15 +6,10 @@
 
 struct route_info;
 
-struct peer_features {
-	u8 *localfeatures;
-	u8 *globalfeatures;
-};
-
 struct gossip_getnodes_entry {
 	struct node_id nodeid;
 	s64 last_timestamp; /* -1 means never: following fields ignored */
-	u8 *globalfeatures;
+	u8 *features;
 	struct wireaddr *addresses;
 	u8 alias[32];
 	u8 color[3];
@@ -44,10 +39,6 @@ struct gossip_getnodes_entry *
 fromwire_gossip_getnodes_entry(const tal_t *ctx, const u8 **pptr, size_t *max);
 void towire_gossip_getnodes_entry(u8 **pptr,
 				  const struct gossip_getnodes_entry *entry);
-
-struct peer_features *
-fromwire_peer_features(const tal_t *ctx, const u8 **pptr, size_t *max);
-void towire_peer_features(u8 **pptr, const struct peer_features *features);
 
 void fromwire_route_hop(const u8 **pprt, size_t *max, struct route_hop *entry);
 void towire_route_hop(u8 **pprt, const struct route_hop *entry);